+++ CHANGELOG.md
@@ -10,9 +10,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates `Tooltip` component with `placement` and `offset` props for custom positioning.
 - Updates the `Table` component to use the `isScrollable` prop to enable horizontal scrolling.
 - Updates the `Table` component to use the `customStyles` prop to allow for custom column styles (i.e. width, text alignment, etc.).
 - Updates the `Table` component to use the `tableTextSize` prop to set the size of the text within the table.
+
+### Fixes
+
+- Fixes a duplicate `id` issue in the `Image` component. This happened when an aspect ratio value rendered a wrapper div with the same `id` as the `img` element. This was not picked up by internal accessibility tests but in a consuming application.
 
 ## 3.3.0 (August 29, 2024)