+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 3.3.1 (September 5, 2024)
+
 ### Updates
 
 - Updates `Tooltip` component with `placement` and `offset` props for custom positioning.